Dynamic dashboard pie chart done. Key Decisions: 1) Local analysis, with analysis being done, whenever refresh button pressed. Then the Summary data is saved in the DB.
@ToDo
1) Add refresh button to pie chart
2) Fix bar chart on second dashboard

+++ mobisens_ui/src/edu/cmu/sv/lifelogger/BarChartActivityLevel.java
@@ -303,9 +303,9 @@ public class BarChartActivityLevel extends Activity {
 			return row;
 		}
 		
-		public int calculateWidthOfBar(int percentage){
+		public int calculateWidthOfBar(double percentage){
 			/* Find the maximum value of all the percentages for the activities */
-			int maxVal = 0;
+			double maxVal = 0;
 			for (ActivityItem a: dataList){
 				if(a.getmPercentage() > maxVal)
 					maxVal = a.getmPercentage();
@@ -314,7 +314,7 @@ public class BarChartActivityLevel extends Activity {
 			/* Extrapolate the percentage to fit maximum width */
 			
 			
-			return 200*percentage/maxVal ;
+			return (int) ((int) 200*percentage/maxVal) ;
 			
 			
 		}
